+++ src/App/DataFixtures/CustomerFixtures.php
@@ -1,0 +1,50 @@
+<?php
+
+/*
+ * This file is part of API as a Service.
+ *
+ * Copyright (c) 2019 Christian Siewert <christian@sieware.international>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\DataFixtures;
+
+use App\Entity\Customer;
+use Doctrine\Persistence\ObjectManager;
+use Doctrine\Bundle\FixturesBundle\Fixture;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+
+/**
+ * @author Christian Siewert <christian@sieware.international>
+ */
+class CustomerFixtures extends Fixture
+{
+    /**
+     * @var UserPasswordEncoderInterface
+     */
+    private $encoder;
+
+    /**
+     * @param UserPasswordEncoderInterface $encoder
+     */
+    public function __construct(UserPasswordEncoderInterface $encoder)
+    {
+        $this->encoder = $encoder;
+    }
+
+    /**
+     * @param ObjectManager $manager
+     */
+    public function load(ObjectManager $manager)
+    {
+        $customer = new Customer();
+        $customer->setEmail('test.user@aaas.api');
+        $customer->setRoles(['ROLE_AAAS_USER']);
+        $customer->setPassword($this->encoder->encodePassword($customer, 'test'));
+
+        $manager->persist($customer);
+        $manager->flush();
+    }
+}